Removing printf warnings, fixing multiple targets

+++ apps/er-coap/er-coap-res-authz-info.c
@@ -37,7 +37,7 @@ static void res_post_handler(void *request, void *response, uint8_t *buffer, uin
 
   printf("Received new token to be stored.\n");
   size_t token_len = REST.get_request_payload(request, (const uint8_t **)&cbor_token);
-  printf("token_len is %ld\n", token_len);
+  printf("token_len is %u\n", (unsigned int) token_len);
   REST.set_header_content_type(response, REST.type.TEXT_PLAIN);
 
   if(token_len == 0) {
